fix: Fix runtime warnings and errors of numpy and torch usages

+++ cosypose/evaluation/meters/utils.py
@@ -22,7 +22,7 @@ def add_inst_num(infos,
                  group_keys=['scene_id', 'view_id', 'label'],
                  key='pred_inst_num'):
 
-    inst_num = np.empty(len(infos), dtype=np.int)
+    inst_num = np.empty(len(infos), dtype=np.int32)
     for group_name, group_ids in infos.groupby(group_keys).groups.items():
         inst_num[group_ids.values] = np.arange(len(group_ids))
     infos[key] = inst_num
